Refactor metrics module structure
- Move build_summary out of html_printer (it powered all three
  output formats) into a dedicated metrics/summary.py.
- Move stale-PR aggregation (_calculate_summary) into
  calculator.py as summarize_stale_prs, next to get_stale_prs.
- Drop the duplicate sort in the stale printer; get_stale_prs
  already orders by age.
- Co-locate the per-target printers under metrics/printer/:
  dev_printer.py -> printer/dev.py
  repo_printer.py -> printer/repo.py
  html_printer.py -> printer/html.py
  stale_printer.py -> printer/stale.py
  Public re-exports in metrics/__init__.py and
  metrics/printer/__init__.py preserve external imports.
