+++ git_dev_metrics/metrics/printer/repo.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from .health import calculate_health_score
+from ..health import calculate_health_score
 
 REPO_COLUMNS = [
     "Repo",
